pyproject.toml
[tool.poetry]
name = "learningdjango"
version = "0.1.0"
description = ""
authors = ["SaugatMgr <saugatmagar433@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
django-jazzmin = "^2.6.0"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.2.0"
pyotp = "^2.9.0"
celery = {extras = ["redis"], version = "^5.3.6"}
django-celery-beat = "^2.6.0"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
